Split Marpit parse image plugin into 2 plugins
